Swap dfu-util git url
I got an error when trying to install dfu-util. It said gitorious
wasn't up.

I went there in a browser, and it looks like they're pushing their
repositories to the Internet Archive project to be in read-only
mode. I swapped out the git url from gitorious.org to the one
listed on the dfu-util page.

Afterward, everything built fine.